+++ chalicelib/covid19bot.py
@@ -4,16 +4,17 @@ from telegram import Bot, Update
 from telegram.ext import Dispatcher, CommandHandler
 from telegram import ParseMode
 
-from chalicelib.ddb_manage import (
-    get_last_record_by_region,
-    DDBUnknownRegion,
-    DDBNoCases
+from chalicelib.sns_manage import publish_to_sns
+from chalicelib.models import Subscription
+from chalicelib.messages import (
+    TGCaseStatsMessage,
+    TGHelpMessage,
+    TGSubscribeMessage
 )
 from chalicelib.configer import (
     APP_NAME,
     TG_TOKEN,
-    DDB_TOTAL_TABLE_NAME,
-    DDB_CASES_TABLE_NAME
+    SNS_TOPIC_FOR_SUBSCRIPTIONS
 )
 
 
@@ -29,50 +30,68 @@ def get_update(request):
 
 def start(update, context):
     """Send a message when the command /start is issued."""
-    update.message.reply_text('hello eugene!')
+    message = 'Hello! I can provide you COVID-19 statistics by a country'
+    update.message.reply_text(
+        message,
+        parse_mode=ParseMode.MARKDOWN
+    )
 
 
 def help(update, context):
-    """Send a message when the command /help is issued."""
+    """Send a message when the command /help"""
+    message = TGHelpMessage().get_message()
     update.message.reply_text(
-        'Hello! I can provide statistics about COVID-19 by regions.\n'
-        'Commands:\n'
-        '/russia - COVID-19 statistics Russia'
+        message,
+        parse_mode=ParseMode.MARKDOWN
+    )
+
+
+def subscribe(update, context):
+    """Mock to subscribe command"""
+    message = 'You must specify a country to get a subscription. Example: ' \
+        '/subscribe\\_russia'
+    update.message.reply_text(
+        message,
+        parse_mode=ParseMode.MARKDOWN
+    )
+
+
+def subscribe_russia(update, context):
+    """Get Russia stats subscription"""
+    region, chat_id = 'russia', update['message']['chat']['id']
+    publish_to_sns(Subscription(chat_id, region), SNS_TOPIC_FOR_SUBSCRIPTIONS)
+    message = TGSubscribeMessage(region).get_message()
+    update.message.reply_text(
+        message,
+        parse_mode=ParseMode.MARKDOWN
+    )
+
+
+def unsubscribe_russia(update, context):
+    """Delete Russia subscription"""
+    update.message.reply_text(
+        "It'll be implemented soon",
+        parse_mode=ParseMode.MARKDOWN
+    )
+
+
+def unsubscribe_all(update, context):
+    """Delete all subscriptions"""
+    update.message.reply_text(
+        "It'll be implemented soon",
+        parse_mode=ParseMode.MARKDOWN
     )
 
 
 def russia(update, context):
-    """Send rus stats"""
-    # TODO: replace to template class
+    """Send Russia stats"""
     lg.info('Russia handler is calling...')
-    template = f'**Russia: Covid Statistics**\n\n'
-    total_header = f'*Total cases*\n'
-    new_header = f'*New cases*\n'
-    try:
-        total = get_last_record_by_region('russia', DDB_TOTAL_TABLE_NAME)
-        lg.debug(f'Total is {total}')
-        tot_msg = f'On *{total["actual_date"]}*\n' \
-                  f'   __Total cases__: {total["case_total"]:,}\n' \
-                  f'   __Recovered__: {total["recovered_total"]:,}\n' \
-                  f'   __Left__: {total["death_total"]:,}\n\n'
-    except (DDBUnknownRegion, DDBNoCases) as e:
-        tot_msg = f'{getattr(e, "message", repr(e))}\n'
-    except Exception:
-        tot_msg = 'Internal error, try again later, please\n'
-    try:
-        new = get_last_record_by_region('russia', DDB_CASES_TABLE_NAME)
-        lg.debug(f'New is {new}')
-        new_msg = f'On *{new["actual_date"]}*\n' \
-                  f'   __New cases__: {new["case_new"]:,}\n' \
-                  f'   __New recovered__: {new["recovered_new"]:,}\n' \
-                  f'   __Left__: {new["death_new"]:,}'
-    except (DDBUnknownRegion, DDBNoCases) as e:
-        new_msg = f'{getattr(e, "message", repr(e))}\n'
-    except Exception:
-        new_msg = 'Internal error, try again later, please\n'
-    message = template + total_header + tot_msg + new_header + new_msg
+    message = TGCaseStatsMessage('russia').get_message()
     lg.debug(f'Output message is {message}')
-    update.message.reply_text(message, parse_mode=ParseMode.MARKDOWN)
+    update.message.reply_text(
+        message,
+        parse_mode=ParseMode.MARKDOWN
+    )
 
 
 def add_handlers(dispatcher):
@@ -80,6 +99,13 @@ def add_handlers(dispatcher):
     dispatcher.add_handler(CommandHandler("start", start))
     dispatcher.add_handler(CommandHandler("help", help))
     dispatcher.add_handler(CommandHandler("russia", russia))
+    dispatcher.add_handler(CommandHandler("subscribe", subscribe))
+    dispatcher.add_handler(CommandHandler(
+        "subscribe_russia", subscribe_russia))
+    dispatcher.add_handler(CommandHandler(
+        "unsubscribe_russia", unsubscribe_russia))
+    dispatcher.add_handler(CommandHandler(
+        "unsubscribe_all", unsubscribe_all))
 
 
 add_handlers(dispatcher)