+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+[0.0.5] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+- implemented a subscription flow
+- implemented a Telegram notifications flow
+- extended aws schema
+- combined Covid total and case models (changed primary and sort keys)
+
 [0.0.4] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 - added create-aws-schema for simple local and AWS deploying 
 - implemented blueprints for views